Create chat controller tests

+++ app/controllers/chat_rooms_controller.rb
@@ -3,22 +3,26 @@ class ChatRoomsController < ApplicationController
 
   def create
     invitation = Invitation.find_by(id: params[:invitation][:id])
-    @chat_room = ChatRoom.create(creator_id: invitation.sender.id , invitee_id: invitation.recipient.id)
-    @first_chat = Chat.create(student_id: invitation.recipient.id,
-      native_speaker_id: invitation.sender.id,
-      topic_id: invitation.recipient.topics.sample.id, chat_room_id: @chat_room.id)
-    @second_chat = Chat.create(student_id: invitation.sender.id,
-      native_speaker_id: invitation.recipient.id,
-      topic_id: invitation.sender.topics.sample.id, chat_room_id: @chat_room.id)
-    @chat_room.chats << @first_chat << @second_chat
-    Invitation.where(sender_id: invitation.sender.id).delete_all
+    @chat_room = ChatRoom.new(creator_id: invitation.sender.id , invitee_id: invitation.recipient.id)
+    if @chat_room.save
+      @first_chat = Chat.create(student_id: invitation.recipient.id,
+        native_speaker_id: invitation.sender.id,
+        topic_id: invitation.recipient.topics.sample.id, chat_room_id: @chat_room.id)
+      @second_chat = Chat.create(student_id: invitation.sender.id,
+        native_speaker_id: invitation.recipient.id,
+        topic_id: invitation.sender.topics.sample.id, chat_room_id: @chat_room.id)
+      @chat_room.chats << @first_chat << @second_chat
+      Invitation.where(sender_id: invitation.sender.id).delete_all
 
-    channel = 'private-conversation.' + @chat_room.creator_id.to_s
-    data = @chat_room.id
-    Pusher.trigger(channel, 'create_chat_room', data.to_json)
+      channel = 'private-conversation.' + @chat_room.creator_id.to_s
+      data = @chat_room.id
+      Pusher.trigger(channel, 'create_chat_room', data.to_json)
 
-    redirect_to chat_room_path(@chat_room)
-    # @chat_room = ChatRoom.create()
+      redirect_to chat_room_path(@chat_room)
+    else
+      flash[:alert] = ["Unable to create chat room"]
+      redirect_to root_path
+    end
   end
 
   def update
@@ -44,8 +48,8 @@ class ChatRoomsController < ApplicationController
             chats: @chat_room.chats,
             current_user: {
               user: current_user,
-              country: current_user.country,
-              language: current_user.native_language
+              country: current_user.try(:country),
+              language: current_user.try(:native_language)
             },
             other_user: {
               user: @other_user,
